+++ AmpTools/GPUManager/GPUNICalcKernel.cu
@@ -40,8 +40,8 @@
 __global__ void
 ni_calc_kernel( int nElements, double* pdDevNICalc,
                 GDouble* pfDevAmps, GDouble* pfDevWeights,
-                unsigned int startEvent, unsigned int nEvents, 
-                unsigned int nTrueEvents )
+                size_t startEvent, size_t nEvents, 
+                size_t nTrueEvents )
 {
 
   // used shared memory block for amplitude indices and results
@@ -75,11 +75,11 @@ ni_calc_kernel( int nElements, double* pdDevNICalc,
   __syncthreads();
   
   // this is the event index in the chunk
-  unsigned int iEvtChunk = iThread +
+  size_t iEvtChunk = iThread +
             ( blockIdx.x + blockIdx.y * gridDim.x ) * GPU_BLOCK_SIZE_SQ;
 
   // this is the event index in the overall set of events 
-  unsigned int iEvt = iEvtChunk + startEvent;  
+  size_t iEvt = iEvtChunk + startEvent;  
 
   if( iEvt < nTrueEvents ){ // do not compute for the padding events
 
@@ -128,11 +128,11 @@ ni_calc_kernel( int nElements, double* pdDevNICalc,
 
 
 extern "C" void GPU_ExecNICalcKernel( dim3 dimGrid, dim3 dimBlock,
-       	   			                      unsigned int sharedSize,
+       	   			                      size_t sharedSize,
                                       int nElements, double* pdDevNICalc,
                                       GDouble* pfDevAmps, GDouble* pfDevWeights,
-                                      unsigned int startEvent, unsigned int nEvents, 
-                                      unsigned int nTrueEvents, unsigned int maxSize )
+                                      size_t startEvent, size_t nEvents, 
+                                      size_t nTrueEvents, size_t maxSize )
 {
 
    if( maxSize )
